Windows: Powershll 支持 cnb 国内源，默认使用 cnb 国内镜像源。

+++ Windows/PowerShell/按需下载万象方案-词库-模型-utf-8.ps1
@@ -1,6 +1,9 @@
 ############# 自动更新配置项，配置好后将 AutoUpdate 设置为 true 即可 #############
 $AutoUpdate = $false;
 
+# 是否使用 CNB 镜像源，如果设置为 $true，则从 CNB 获取资源；否则从 GitHub 获取。
+$UseCnbMirrorSource = $true
+
 # 设置自动更新时，是否更新方案、词库、模型，不想更新某项就改成false
 $IsUpdateSchemaDown = $true
 $IsUpdateDictDown = $true
@@ -27,6 +30,8 @@ $InputSchemaType = "7";
 
 ############# 自动更新配置项，配置好后将 AutoUpdate 设置为 true 即可 #############
 
+$Debug = $false;
+
 $UpdateToolsVersion = "DEFAULT_UPDATE_TOOLS_VERSION_TAG";
 if ($UpdateToolsVersion.StartsWith("DEFAULT")) {
     Write-Host "您下载的是非发行版脚本，请勿直接使用，请去 releases 页面下载最新版本：https://github.com/expoli/rime-wanxiang-update-tools/releases" -ForegroundColor Yellow;
@@ -35,23 +40,31 @@ if ($UpdateToolsVersion.StartsWith("DEFAULT")) {
 }
 
 # 设置仓库所有者和名称
-$SchemaOwner = "amzxyz"
-$SchemaRepo = "rime_wanxiang"
-$GramRepo = "RIME-LMDG"
-$GramReleaseTag = "LTS"
-$GramModelFileName = "wanxiang-lts-zh-hans.gram"
-$ReleaseTimeRecordFile = "release_time_record.json"
 $UpdateToolsOwner = "expoli"
 $UpdateToolsRepo = "rime-wanxiang-update-tools"
 # 定义临时文件路径
 $tempSchemaZip = Join-Path $env:TEMP "wanxiang_schema_temp.zip"
 $tempDictZip = Join-Path $env:TEMP "wanxiang_dict_temp.zip"
 $tempGram = Join-Path $env:TEMP "wanxiang-lts-zh-hans.gram"
-$tempGramMd5 = Join-Path $env:TEMP "wanxiang-lts-zh-hans.gram.md5"
 $SchemaExtractPath = Join-Path $env:TEMP "wanxiang_schema_extract"
 $DictExtractPath = Join-Path $env:TEMP "wanxiang_dict_extract"
 
-$Debug = $false;
+$GramModelFileName = "wanxiang-lts-zh-hans.gram"
+$ReleaseTimeRecordFile = "release_time_record.json"
+
+if ($UseCnbMirrorSource) {
+    $SchemaOwner = "amzxyz"
+    $SchemaRepo = "rime-wanxiang"
+    $GramRepo = "RIME-LMDG"
+    $GramReleaseTag = "model"
+    $DictReleaseTag = "v1.0.0"
+} else {
+    $SchemaOwner = "amzxyz"
+    $SchemaRepo = "rime_wanxiang"
+    $GramRepo = "RIME-LMDG"
+    $GramReleaseTag = "LTS"
+    $DictReleaseTag = "dict-nightly"
+}
 
 $KeyTable = @{
     "0" = "base";
@@ -202,11 +215,14 @@ function Start-WeaselServer {
 
 function Start-WeaselReDeploy{
     $defaultShortcutPath = "C:\ProgramData\Microsoft\Windows\Start Menu\Programs\小狼毫输入法\【小狼毫】重新部署.lnk"
+    $backupEnglishShortcutPath = "C:\ProgramData\Microsoft\Windows\Start Menu\Programs\Weasel\Weasel Deploy.lnk"
     if (Test-Path -Path $defaultShortcutPath) {
         Write-Host "找到默认【小狼毫】重新部署快捷方式，将执行" -ForegroundColor Green
         Invoke-Item -Path $defaultShortcutPath
-    }
-    else {
+    } elseif (Test-Path -Path $backupEnglishShortcutPath) {
+        Write-Host "找到默认【小狼毫】重新部署快捷方式，将执行" -ForegroundColor Green
+        Invoke-Item -Path $backupEnglishShortcutPath
+    } else {
         Write-Host "未找到默认的【小狼毫】重新部署快捷方式，将尝试执行默认的重新部署命令" -ForegroundColor Yellow
         Write-Host "跳过触发重新部署" -ForegroundColor Yellow
     }
@@ -227,7 +243,12 @@ function Test-VersionSuffix {
     )
     # tag_name = v1.0.0 or v1.0
     $pattern = 'v(\d+)(\.\d+)+'
-    return $url -match $pattern
+    if ($UseCnbMirrorSource) {
+        return $url -match $pattern | Where-Object { $_ -notmatch $DictReleaseTag -and $_ -notmatch $GramReleaseTag }
+    }
+    else {
+        return $url -match $pattern
+    }
 }
 
 function Test-DictSuffix {
@@ -235,12 +256,59 @@ function Test-DictSuffix {
         [string]$url
     )
 
-    # tag_name = dict-nightly
-    $pattern = 'dict-nightly'
-    return $url -match $pattern
+    return $url -match $DictReleaseTag
 }
 
-function Get-ReleaseInfo {
+function Test-CnbGramSuffix {
+    param(
+        [string]$url
+    )
+    # tag_name = model
+    return $url -match $GramReleaseTag
+}
+
+function Get-CnbReleaseInfo {
+    param(
+        [string]$owner,
+        [string]$repo
+    )
+
+    $apiUrl = "https://cnb.cool/$owner/$repo/-/releases"
+
+    try {
+        Write-Host "正在从 CNB 页面获取信息: $apiUrl" -ForegroundColor Cyan
+        $htmlContent = Invoke-WebRequest -Uri $apiUrl -UseBasicParsing | Select-Object -ExpandProperty Content
+      
+        # 使用正则表达式匹配<script id="__NEXT_DATA__">标签内的内容
+        $regex = '(?s)<script id="__NEXT_DATA__"[^>]*>(.*?)</script>'
+        if ($htmlContent -match $regex) {
+            $jsonData = $matches[1]
+            $jsonDataFormat = $jsonData | ConvertFrom-Json
+            if ($jsonDataFormat.props.pageProps.initialState.slug.repo.releases.list.data.releases){
+                $releaseData = $jsonDataFormat.props.pageProps.initialState.slug.repo.releases.list.data.releases
+                Write-Host "成功获取 CNB release 版本信息" -ForegroundColor Green
+                if ($releaseData.assets.Count -eq 0) {
+                    Write-Error "CNB release 版本没有可下载资源"
+                    Exit-Tip 1
+                }
+                return $releaseData
+            } else {
+                Write-Error "错误：在页面中未找到 'release' 数据。"
+                Exit-Tip 1
+            }
+        } else {
+            Write-Error "错误：在页面中未找到 '__NEXT_DATA__' JSON 数据。"
+            Exit-Tip 1
+        }
+    }
+    catch {
+        Write-Error "错误：下载或解析CNB页面失败: $pageUrl"
+        Write-Error $_.Exception.Message
+        Exit-Tip 1
+    }
+}
+
+function Get-GithubReleaseInfo {
     param(
         [string]$owner,
         [string]$repo
@@ -280,7 +348,24 @@ function Get-ReleaseInfo {
     return $response
 }
 
-$UpdateTollsResponse = Get-ReleaseInfo -owner $UpdateToolsOwner -repo $UpdateToolsRepo
+function Get-ReleaseInfo {
+    param(
+        [string]$owner,
+        [string]$repo,
+        [bool]$updateToolFlag = $false
+    )
+    # 构建API请求URL
+    if ($updateToolFlag) {
+        return Get-GithubReleaseInfo -owner $owner -repo $repo
+    }
+    if ($UseCnbMirrorSource){
+        return Get-CnbReleaseInfo -owner $owner -repo $repo
+    } else {
+        return Get-GithubReleaseInfo -owner $owner -repo $repo
+    }
+}
+
+$UpdateTollsResponse = Get-ReleaseInfo -owner $UpdateToolsOwner -repo $UpdateToolsRepo -updateToolFlag $true
 
 # 检查是否有新版本,如果获取的版本信息比现在的版本信息(UpdateToolsVersion)新，则提示用户更新
 # 版本格式:v3.4.0,v3.4.1,v3.4.1-rc1,不比较 rc 版本,
@@ -289,8 +374,8 @@ if ($UpdateTollsResponse.Count -eq 0) {
     Write-Host "没有找到更新工具的版本信息，请检查网络连接或仓库是否存在" -ForegroundColor Red
     Exit-Tip 1
 }
-# 过滤掉包含 -rc 的 tag_name
-$StableUpdateToolsReleases = $UpdateTollsResponse | Where-Object { $_.tag_name -notmatch '-rc' }
+
+$StableUpdateToolsReleases = $UpdateTollsResponse
 if ($StableUpdateToolsReleases.Count -eq 0) {
     Write-Host "没有找到稳定版的更新工具版本信息" -ForegroundColor Yellow
 } else {
@@ -305,17 +390,36 @@ if ($StableUpdateToolsReleases.Count -eq 0) {
 
 # 获取最新的版本信息
 $SchemaResponse = Get-ReleaseInfo -owner $SchemaOwner -repo $SchemaRepo
-$GramResponse = Get-ReleaseInfo -owner $SchemaOwner -repo $GramRepo
+if ($UseCnbMirrorSource) {
+    $GramResponse = $SchemaResponse
+} else {
+    $GramResponse = Get-ReleaseInfo -owner $SchemaOwner -repo $GramRepo
+}
 
 $SelectedDictRelease = $null
 $SelectedSchemaRelease = $null
 $SelectedGramRelease = $null
 
-foreach ($release in $SchemaResponse) {
-    if (($null -eq $SelectedDictRelease) -and (Test-DictSuffix -url $release.tag_name)) {
-        $SelectedDictRelease = $release
+function Get-ReleaseTagName {
+    param(
+        [object]$release
+    )
+    if ($UseCnbMirrorSource) {
+        $tag_name = $release.tagRef
+    } else {
+        $tag_name = $release.tag_name
     }
-    if ((Test-VersionSuffix -url $release.tag_name) -and (-not $SelectedSchemaRelease)) {
+    return $tag_name
+}
+
+foreach ($release in $SchemaResponse) {
+    $tag_name = Get-ReleaseTagName -release $release
+
+    if (($null -eq $SelectedDictRelease) -and (Test-DictSuffix -url $tag_name)) {
+        $SelectedDictRelease = $release
+        continue
+    }
+    if ((-not $SelectedSchemaRelease) -and (Test-VersionSuffix -url $tag_name)) {
         $SelectedSchemaRelease = $release
     }
     if ($SelectedDictRelease -and $SelectedSchemaRelease) {
@@ -324,30 +428,38 @@ foreach ($release in $SchemaResponse) {
 }
 
 foreach ($release in $GramResponse) {
+    $tag_name = Get-ReleaseTagName -release $release
     if ($Debug) {
-        Write-Host "release.tag_name: $($release.tag_name)" -ForegroundColor Green
+        Write-Host "release.tag_name: $tag_name" -ForegroundColor Green
         Write-Host "GramReleaseTag: $GramReleaseTag" -ForegroundColor Green
     }
-    if ($release.tag_name -eq $GramReleaseTag) {
+    if ($tag_name -match $GramReleaseTag) {
         $SelectedGramRelease = $release
     }
 }
 
 if ($SelectedDictRelease -and $SelectedSchemaRelease -and $SelectedGramRelease) {
-    Write-Host "解析出最新的词库链接为：$($SelectedDictRelease.html_url)" -ForegroundColor Green
-    Write-Host "解析出最新的版本链接为：$($SelectedSchemaRelease.html_url)" -ForegroundColor Green
-    Write-Host "解析出最新的模型链接为：$($SelectedGramRelease.html_url)" -ForegroundColor Green
+    if (-not $UseCnbMirrorSource) {
+        Write-Host "解析出最新的词库链接为：$($SelectedDictRelease.html_url)" -ForegroundColor Green
+        Write-Host "解析出最新的版本链接为：$($SelectedSchemaRelease.html_url)" -ForegroundColor Green
+        Write-Host "解析出最新的模型链接为：$($SelectedGramRelease.html_url)" -ForegroundColor Green
+    }
 } else {
     Write-Error "未找到符合条件的版本或词库链接"
     Exit-Tip 1
 }
 
 # 获取最新的版本的tag_name
-Write-Host "方案最新的版本为：$($SelectedSchemaRelease.tag_name)"
-Write-Host "方案更新日志: " -ForegroundColor Yellow
-Write-Host $SelectedSchemaRelease.body -ForegroundColor Yellow
+if (-not $UseCnbMirrorSource) {
+    Write-Host "方案最新的版本为：$($SelectedSchemaRelease.tag_name)"
+    Write-Host "方案更新日志: " -ForegroundColor Yellow
+    Write-Host $SelectedSchemaRelease.body -ForegroundColor Yellow
+} else {
+    Write-Host "方案最新的版本为：$($SelectedSchemaRelease.tagRef)"
+    Write-Host "方案更新日志: " -ForegroundColor Yellow
+    Write-Host $SelectedSchemaRelease.body -ForegroundColor Yellow
+}
 
-$SchemaTag = $SelectedSchemaRelease.tag_name
 
 $promptSchemaType = "请选择你要下载的方案类型的编号: `n$SchemaDownloadTip"
 $promptAllUpdate = "是否更新所有内容（方案、词库、模型）:`n[0]-更新所有; [1]-不更新所有"
@@ -489,7 +601,7 @@ function Save-TimeRecord {
             Write-Host "警告：无法读取时间记录文件，将创建新的记录" -ForegroundColor Yellow
         }
     }
-
+0
     $timeData[$key] = $value
     
     try {
@@ -628,15 +740,22 @@ function Download-Files {
     )
     
     try {
-        $downloadUrl = $assetInfo.browser_download_url
+        if ($UseCnbMirrorSource) {
+            $downloadUrl = "https://cnb.cool" + $assetInfo.path
+        } else {
+            $downloadUrl = $assetInfo.browser_download_url
+        }
+        
         Write-Host "正在下载文件:$($assetInfo.name)..." -ForegroundColor Green
         Invoke-WebRequest -Uri $downloadUrl -OutFile $outFilePath -UseBasicParsing
         Write-Host "下载完成" -ForegroundColor Green
-        $SHA256 = $assetInfo.digest.Split(":")[1]
-        if (-not (Test-FileSHA256 -FilePath $outFilePath -CompareSHA256 $SHA256)) {
-            Write-Host "SHA256 校验失败，删除文件" -ForegroundColor Red
-            Remove-Item -Path $outFilePath -Force
-            Exit-Tip 1
+        if (-not $UseCnbMirrorSource) {
+            $SHA256 = $assetInfo.digest.Split(":")[1]
+            if (-not (Test-FileSHA256 -FilePath $outFilePath -CompareSHA256 $SHA256)) {
+                Write-Host "SHA256 校验失败，删除文件" -ForegroundColor Red
+                Remove-Item -Path $outFilePath -Force
+                Exit-Tip 1
+            }
         }
     }
     catch {
@@ -674,13 +793,24 @@ if ($InputSchemaDown -eq "0" -or $InputDictDown -eq "0" -or $InputGramModel -eq 
     Exit-Tip 0
 }
 
+function Get-UpdateAtObj {
+    param (
+        [object]$assetInfo
+    )
+    if ($UseCnbMirrorSource) {
+        return $assetInfo.updatedAt
+    } else {
+        return $assetInfo.updated_at
+    }
+}
+
 $UpdateFlag = $false
 
 if ($InputSchemaDown -eq "0") {
     # 下载方案
     $SchemaUpdateTimeKey = $KeyTable[$InputSchemaType] + "_schema_update_time"
     $SchemaUpdateTime = Get-TimeRecord -filePath $TimeRecordFile -key $SchemaUpdateTimeKey
-    $SchemaRemoteTime = [datetime]::Parse($ExpectedSchemaTypeInfo.updated_at)
+    $SchemaRemoteTime = [datetime]::Parse($(Get-UpdateAtObj -assetInfo $ExpectedSchemaTypeInfo))
     Write-Host "正在检查方案是否需要更新..." -ForegroundColor Yellow
     Write-Host "本地时间: $SchemaUpdateTime" -ForegroundColor Green
     Write-Host "远程时间: $SchemaRemoteTime" -ForegroundColor Green
@@ -740,7 +870,7 @@ if ($InputDictDown -eq "0") {
     # 下载词库
     $DictUpdateTimeKey = $KeyTable[$InputSchemaType] + "_dict_update_time"
     $DictUpdateTime = Get-TimeRecord -filePath $TimeRecordFile -key $DictUpdateTimeKey
-    $DictRemoteTime = [datetime]::Parse($ExpectedDictTypeInfo.updated_at)
+    $DictRemoteTime = [datetime]::Parse($(Get-UpdateAtObj -assetInfo $ExpectedDictTypeInfo))
     Write-Host "正在检查词库是否需要更新..." -ForegroundColor Yellow
     Write-Host "本地时间: $DictUpdateTime" -ForegroundColor Green
     Write-Host "远程时间: $DictRemoteTime" -ForegroundColor Green
@@ -800,7 +930,7 @@ if ($InputGramModel -eq "0") {
     # 下载模型
     $GramUpdateTimeKey = $GramReleaseTag + "_gram_update_time"
     $GramUpdateTime = Get-TimeRecord -filePath $TimeRecordFile -key $GramUpdateTimeKey
-    $GramRemoteTime = [datetime]::Parse($ExpectedGramTypeInfo.updated_at)
+    $GramRemoteTime = [datetime]::Parse($(Get-UpdateAtObj -assetInfo $ExpectedGramTypeInfo))
     Write-Host "正在检查模型是否需要更新..." -ForegroundColor Yellow
     # 检查目标文件 $targetDir/$tempGram 是否存在
     $filePath = Join-Path $targetDir $GramModelFileName
@@ -812,7 +942,10 @@ if ($InputGramModel -eq "0") {
     if (Compare-UpdateTime -localTime $GramUpdateTime -remoteTime $GramRemoteTime) {
         Update-GramModel
         $UpdateFlag = $true
-    }elseif (Test-Path -Path $filePath) {
+    } elseif (Test-Path -Path $filePath) {
+        if ($UseCnbMirrorSource) {
+            return
+        }
         # 计算目标文件的MD5
         $localSHA256 = (Get-FileHash $filePath -Algorithm SHA256).Hash.ToLower()
         # 计算远程文件的MD5
@@ -822,7 +955,7 @@ if ($InputGramModel -eq "0") {
             Write-Host "模型MD5不匹配，需要更新" -ForegroundColor Red
             Update-GramModel
             $UpdateFlag = $true
-        }   
+        }
     } else {
         Write-Host "模型不存在，需要更新" -ForegroundColor Red
         Update-GramModel
